requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_community
sentence-transformers
faiss-cpu
beautifulsoup4
pydantic
langchain-ollama
